pasar id a string porque en mongo son strings

+++ chotuve/app/src/main/java/com/taller2/chotuve/vista/ver_video/VerVideoActivity.kt
@@ -57,7 +57,7 @@ class VerVideoActivity: AppCompatActivity(), VistaVerVideo {
                 ponerReproductorEnPantallaCompleta()
             }
         }
-        val id = intent.getStringExtra(ID_KEY)!!.toLong()
+        val id = intent.getStringExtra(ID_KEY)!!
         presentador.obtenerVideo(id)
     }
